build: ensure go generate doesn't impact linux builds [#11]
previously, goversioninfo would generate one file named resource.syso,
which the go compiler (linker?) would pick up when building for any OS
and architecture. This caused a build error when targeting Linux because
the Windows-only .syso file was invalid. Adding the -platform-specific
flag ensures the generated files are named so that they're only picked
up when building for Windows.